// only-doc-vibe-main/src/shared/api/cms/cms-schema.ts
export interface components {
  schemas: {
// ...
    ContactUs: {
      id: number;
      documentId: string;
      seo: components["schemas"]["SeoSeoComponent"];
      title: string;
      subtitle: string;
      contact_info_items: components["schemas"]["CommonInfoItemComponent"][];
      form_title: string;
      form_name_placeholder: string;
      form_email_placeholder: string;
      form_message_placeholder: string;
      form_submit_button_text: string;
      submit_popup: string;
      required_field: string;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
      /** Format: date-time */
      publishedAt: string;
      createdBy: {
        id: number;
        documentId: string;
      };
      updatedBy: {
        id: number;
        documentId: string;
      };
      locale: string;
      localizations: {
          id: number;
          documentId: string;
          seo: components["schemas"]["SeoSeoComponent"];
          title: string;
          subtitle: string;
          contact_info_items: components["schemas"]["CommonInfoItemComponent"][];
          form_title: string;
          form_name_placeholder: string;
          form_email_placeholder: string;
          form_message_placeholder: string;
          form_submit_button_text: string;
          submit_popup: string;
          required_field: string;
          /** Format: date-time */
          createdAt: string;
          /** Format: date-time */
          updatedAt: string;
          /** Format: date-time */
          publishedAt: string;
          createdBy: {
            id: number;
            documentId: string;
          };
          updatedBy: {
            id: number;
            documentId: string;
          };
          locale: string;
          localizations: {
              id: number;
              documentId: string;
            }[];
        }[];
    };
// ...
  };
// ...
}