// only-doc-vibe-main/src/shared/constants/page-type.ts
export enum EPageType {
  HOME = "home",
  SERVICE = "service",
  OTHER = "other",
  CONTACT_US = "contact-us",
  ABOUT_US = "about-us",
  LEGAL = "legal",
  LOGIN = "login",
  SIGN_UP = "sign-up",
  BLOG_HOME = "blog-home",
  BLOG_CATEGORY = "blog-category",
  BLOG_ARTICLE = "blog-article",
  BLOG_AUTHOR = "blog-author",
  TEMPLATES_HOME = "templates-home",
  TEMPLATE_CATEGORY = "template-category",
  TEMPLATE_SUBCATEGORY = "template-subcategory",
  TEMPLATE_PAGE = "template-page",
}